requirements.txt
flask
python-dotenv
psycopg2
flask-sqlalchemy
flask-cors
pytest
requests